1)CBOSuite was moved to StatisticsCollectionSuite 2)Reduced number of changed lines in FilterEstomation.evaluateInSet
