+++ sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/statsEstimation/FilterEstimationSuite.scala
@@ -360,9 +360,9 @@ class FilterEstimationSuite extends StatsEstimationTestBase {
   test("evaluateInSet with all zeros") {
     validateEstimatedStats(
       Filter(InSet(attrString, Set(3, 4, 5)),
-        StatsTestPlan(Seq(attrString), 10,
+        StatsTestPlan(Seq(attrString), 0,
           AttributeMap(Seq(attrString ->
-            ColumnStat(distinctCount = Some(0), min = Some(0), max = Some(0),
+            ColumnStat(distinctCount = Some(0), min = None, max = None,
               nullCount = Some(0), avgLen = Some(0), maxLen = Some(0)))))),
       Seq(attrString -> ColumnStat(distinctCount = Some(0))),
       expectedRowCount = 0)
